\name{RootDepth-package}
\alias{RootDepth-package}
\alias{RootDepth}
\docType{package}
\title{
RootDepth: Plant root depth statistics, shape summaries (D95, beta, etc), and plots
}
\description{
RootDepth takes root length, volume, or mass data coded by factors, like genotype and soil properties, and outputs 
}
\details{
\tabular{ll}{
Package: \tab RootDepth\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-17\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Eric Nord, Larry York
Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
